src/requirements.txt
azure.core
streamlit
azure-storage-blob
python-dotenv
azure-cognitiveservices-vision-customvision
